4-tools/requirements.txt
langchain==1.2.17
langchain-openai==1.2.1
python-dotenv==0.19.2
requests==2.32.3
